+++ migrations/2020-02-05-155615_create_users/up.sql
@@ -1,9 +1,16 @@
 -- Your SQL goes here
 CREATE TABLE users (
-  id INT PRIMARY KEY,
-  username VARCHAR(150) NOT NULL,
+  id INT UNSIGNED NOT NULL AUTO_INCREMENT,
+  email VARCHAR(150) NOT NULL,
   password VARCHAR(150) NOT NULL,
-  email VARCHAR(150),
-  firstname VARCHAR(150),
-  lastname VARCHAR(150)
+  firstname VARCHAR(150) NOT NULL,
+  lastname VARCHAR(150) NOT NULL,
+  street VARCHAR(150),
+  number INT UNSIGNED,
+  city VARCHAR(150),
+  zipcode INT UNSIGNED,
+  country VARCHAR(150),
+  phone VARCHAR(150),
+  PRIMARY KEY (id),
+  CONSTRAINT unique_email UNIQUE (email)
 )