+++ src/main.rs
@@ -1,5 +1,7 @@
 #![feature(proc_macro_hygiene, decl_macro)]
 
+/* --------------------------- Load Extern Crates -------------------------- */
+
 #[macro_use]
 extern crate rocket;
 #[macro_use]
@@ -10,86 +12,45 @@ extern crate rocket_contrib;
 extern crate diesel;
 extern crate dotenv;
 
+/* ------------------------------ Local Mod -------------------------------- */
+
 mod authentication;
 mod database;
+mod http;
 mod models;
 mod schema;
 
+/* --------------------------- Load Namespaces ----------------------------- */
+
 use std::path::{Path, PathBuf};
 
-use authentication::guards::Auth;
 use database::MyDbConn;
-use models::user::User;
 
 use rocket::http::Cookies;
 use rocket::response::NamedFile;
 
 use rocket_contrib::templates::Template;
 
-use diesel::prelude::*;
-
-#[derive(Serialize)]
-struct TemplateContext {
-    title: &'static str,
-    name: Option<String>,
-    items: Vec<&'static str>,
-    // This key tells handlebars which template is the parent.
-    parent: &'static str,
-}
-
-#[derive(Serialize)]
-struct UserInfoTemplateContext<'a> {
-    title: &'static str,
-    user: &'a User,
-    parent: &'static str,
-}
-
 /* -------------------------------- Routes --------------------------------- */
 
-#[get("/test?<option>")]
-fn optional(option: Option<u32>) -> &'static str {
-    match option {
-        Some(_n) => "Oh ya un nombre",
-        None => "rien à voir ici",
-    }
-}
-
-#[get("/cookies")]
-fn get_cookies(cookies: Cookies) -> String {
-    cookies
-        .iter()
-        .map(|c| format!("name : {:?}, value : {:?}", c.name(), c.value()))
-        .collect::<String>()
-}
-
+/// Serve application entrypoint
 #[get("/")]
 fn index() -> Template {
     Template::render("layout", &())
 }
 
-#[get("/hidden")]
-fn hidden(auth: Auth) -> Template {
-    Template::render(
-        "hidden",
-        &UserInfoTemplateContext {
-            title: "Hidden",
-            user: &auth.user,
-            parent: "layout",
-        },
-    )
-}
+/// Allow some routes to fetch entrypoint when refreshed
+#[get("/<route>", rank = 10)]
+fn dynamic_routing(route: String) -> Option<Template> {
+    let mut allowed_routes = vec!["profile", "notifications", "settings", "about"];
 
-#[get("/hello/<name>")]
-fn hello(name: String) -> Template {
-    Template::render(
-        "index",
-        &TemplateContext {
-            title: "Hello",
-            name: Some(name),
-            items: vec!["One", "Two", "Three"],
-            parent: "layout",
-        },
-    )
+    allowed_routes.append(&mut authentication::routes::allowed_paths());
+
+    if allowed_routes.contains(&&route[..]) {
+        Some(Template::render("layout", &()))
+    } else {
+        None
+    }
 }
 
 /// Serve static files
@@ -98,12 +59,21 @@ fn files(file: PathBuf) -> Option<NamedFile> {
     NamedFile::open(Path::new("static/").join(file)).ok()
 }
 
+/// Display cookies stored
+// TODO : remove me when done testing
+#[get("/dev/cookies")]
+fn get_cookies(cookies: Cookies) -> String {
+    cookies
+        .iter()
+        .map(|c| format!("name : {:?}, value : {:?}", c.name(), c.value()))
+        .collect::<String>()
+}
+
+/* ----------------------------- Launch Rocket ----------------------------- */
+
 fn main() {
     rocket::ignite()
-        .mount(
-            "/",
-            routes![files, index, optional, hello, hidden, get_cookies],
-        )
+        .mount("/", routes![index, dynamic_routing, files, get_cookies])
         .mount("/", authentication::routes::collect())
         .attach(Template::fairing())
         .attach(MyDbConn::fairing())