+++ src/models/mod.rs
@@ -1,1 +1,2 @@
+pub mod quick_response;
 pub mod user;